feat(protocol): add steer/approval routing infrastructure for agent adapters
- ClaudeAdapter now explicitly overrides steer() and respond_approval()
  returning Unsupported with a descriptive message naming the Claude CLI
  one-shot process limitation, replacing silent default-trait fallthrough
- Add running_adapters DashMap to ThreadManager with register/deregister
  and async steer_active_turn / respond_approval_on_turn routing methods
- turn/steer handler now calls steer_active_turn, forwarding instructions
  to the live adapter when registered (graceful no-op otherwise)
- New turn/respond_approval RPC method routes approval decisions to the
  active adapter, returning Unsupported for non-interactive backends
- Add StreamItem::ApprovalRequest and handle it in process_stream_item:
  appends Item::ApprovalRequest to the turn and emits
  Notification::ApprovalRequest to subscribers
- Add adapter_registry field to HarnessServer (default empty); wired in
  run_turn_lifecycle for register/deregister around the turn

Closes #591
