+++ crates/harness-server/src/handlers/thread.rs
@@ -262,7 +262,8 @@ pub async fn turn_steer(
                 .core
                 .server
                 .thread_manager
-                .steer_turn(&thread_id, &turn_id, instruction)
+                .steer_active_turn(&turn_id, instruction)
+                .await
             {
                 Ok(()) => {
                     persist_thread(state, &thread_id).await;
@@ -272,6 +273,25 @@ pub async fn turn_steer(
             }
         }
         None => RpcResponse::error(id, NOT_FOUND, "turn not found in any thread"),
+    }
+}
+
+pub async fn turn_respond_approval(
+    state: &AppState,
+    id: Option<serde_json::Value>,
+    turn_id: harness_core::types::TurnId,
+    request_id: String,
+    decision: harness_core::agent::ApprovalDecision,
+) -> RpcResponse {
+    match state
+        .core
+        .server
+        .thread_manager
+        .respond_approval_on_turn(&turn_id, request_id, decision)
+        .await
+    {
+        Ok(()) => RpcResponse::success(id, serde_json::json!({ "responded": true })),
+        Err(e) => RpcResponse::error(id, INTERNAL_ERROR, e.to_string()),
     }
 }
